+++ packages/store/ts/exports/internal.ts
@@ -1,3 +1,16 @@
 export * from "../common";
 export * from "../getStoreLogs";
 export * from "../flattenStoreLogs";
+
+export * from "../config/v2/input";
+export * from "../config/v2/output";
+export * from "../config/v2/store";
+export * from "../config/v2/generics";
+export * from "../config/v2/defaults";
+export * from "../config/v2/table";
+export * from "../config/v2/tables";
+export * from "../config/v2/tableShorthand";
+export * from "../config/v2/namespace";
+export * from "../config/v2/namespaces";
+export * from "../config/v2/scope";
+export * from "../config/v2/flattenNamespacedTables";