+++ packages/world-module-erc20/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: {
     "mud.config": "mud.config.ts",
-    index: "ts/exports/index.ts",
+    internal: "ts/exports/internal.ts",
   },
   target: "esnext",
   format: ["esm"],