+++ packages/explorer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @latticexyz/explorer
 
+## 2.2.13
+
+### Patch Changes
+
+- 79d273a: The SQL query editor now supports multi-line input.
+- Updated dependencies [dfc2d64]
+  - @latticexyz/store-sync@2.2.13
+  - @latticexyz/schema-type@2.2.13
+  - @latticexyz/store@2.2.13
+  - @latticexyz/world@2.2.13
+  - @latticexyz/store-indexer@2.2.13
+  - @latticexyz/common@2.2.13
+  - @latticexyz/config@2.2.13
+  - @latticexyz/protocol-parser@2.2.13
+
 ## 2.2.12
 
 ### Patch Changes